Move tests in slgext to appropriate places
