+++ tests/test/misc.rs
@@ -40,3 +40,497 @@ fn futures_ambiguity() {
         }
     }
 }
+
+#[test]
+fn basic() {
+    test! {
+        program {
+            trait Sized { }
+
+            struct i32 { }
+            impl Sized for i32 { }
+        }
+
+        goal {
+            forall<T> { if (T: Sized) { T: Sized } }
+        } yields_all[SolverChoice::slg(10, None)] {
+            "substitution [], lifetime constraints []"
+        }
+    }
+}
+
+#[test]
+fn breadth_first() {
+    test! {
+        program {
+            trait Sized { }
+
+            struct i32 { }
+            impl Sized for i32 { }
+
+            struct Vec<T> { }
+            impl<T> Sized for Vec<T> where T: Sized { }
+
+            struct Slice<T> { }
+            impl<T> Sized for Slice<T> where T: Sized { }
+        }
+
+        goal {
+            exists<T> { T: Sized }
+        } yields_first[SolverChoice::slg(10, None)] {
+            "substitution [?0 := i32], lifetime constraints []",
+            "substitution [?0 := Vec<i32>], lifetime constraints []",
+            "substitution [?0 := Slice<i32>], lifetime constraints []",
+            "substitution [?0 := Vec<Vec<i32>>], lifetime constraints []",
+            "substitution [?0 := Slice<Vec<i32>>], lifetime constraints []"
+        }
+    }
+}
+
+/// Make sure we don't get a stack overflow or other badness for this
+/// test from scalexm.
+#[test]
+fn subgoal_abstraction() {
+    test! {
+        program {
+            trait Foo { }
+            struct Box<T> { }
+            impl<T> Foo for T where Box<T>: Foo { }
+        }
+
+        goal {
+            exists<T> { T: Foo }
+        } yields_all[SolverChoice::slg(50, None)] {
+        }
+    }
+}
+
+#[test]
+fn flounder() {
+    test! {
+        program {
+            trait A { }
+
+            struct Vec<T> { }
+            impl<T> A for Vec<T> { }
+        }
+
+        goal {
+            exists<T> { not { T: A } }
+        } yields_first[SolverChoice::slg(10, None)] {
+            "Floundered"
+        }
+    }
+}
+
+// Test that, when solving `?T: Sized`, we only wind up pulling a few
+// answers before we stop.
+// This is similar to the `breadth_first` test, except the order of the
+// FIXME: This is basically the same as `breadth_first`. Is it testing something different?
+#[test]
+fn only_draw_so_many() {
+    test! {
+        program {
+            trait Sized { }
+
+            struct Vec<T> { }
+            impl<T> Sized for Vec<T> where T: Sized { }
+
+            struct i32 { }
+            impl Sized for i32 { }
+
+            struct Slice<T> { }
+            impl<T> Sized for Slice<T> where T: Sized { }
+        }
+
+        goal {
+            exists<T> { T: Sized }
+        } yields_first[SolverChoice::slg(10, None)] {
+            "substitution [?0 := i32], lifetime constraints []",
+            "substitution [?0 := Slice<i32>], lifetime constraints []",
+            "substitution [?0 := Vec<i32>], lifetime constraints []"
+        }
+
+        goal {
+            exists<T> { T: Sized }
+        } yields[SolverChoice::slg(10, Some(2))] {
+            "Ambiguous; no inference guidance"
+        }
+    }
+}
+
+#[test]
+fn only_draw_so_many_blow_up() {
+    test! {
+        program {
+            trait Sized { }
+            trait Foo { }
+
+            struct Vec<T> { }
+            impl<T> Sized for Vec<T> where T: Sized { }
+            impl<T> Foo for Vec<T> where T: Sized { }
+
+            struct i32 { }
+            impl Sized for i32 { }
+
+            struct Slice<T> { }
+            impl<T> Sized for Slice<T> where T: Sized { }
+        }
+
+        goal {
+            exists<T> { T: Foo }
+        } yields[SolverChoice::slg(10, Some(2))] {
+            "Ambiguous; definite substitution for<?U0> { [?0 := Vec<^0>] }"
+        }
+    }
+}
+
+// FIXME: well-formed problems?
+#[test]
+#[ignore]
+fn subgoal_cycle_uninhabited() {
+    test! {
+        program {
+            trait Foo { }
+            struct Box<T> { }
+            struct Vec<T> { }
+            struct u32 { }
+            impl<T> Foo for Box<T> where Box<Vec<T>>: Foo { }
+        }
+
+        // There is no solution here with a finite proof, so we get
+        // back: 0 answer(s) found.
+        goal {
+            exists<T> { T: Foo }
+        } yields_all[SolverChoice::slg(2, None)] {
+            "substitution [], lifetime constraints []"
+        }
+
+        // Unsurprisingly, applying negation succeeds then.
+        goal {
+            not { exists<T> { T: Foo } }
+        } yields_all[SolverChoice::slg(2, None)] {
+            "substitution [], lifetime constraints []"
+        }
+
+        // Equivalent to the previous.
+        goal {
+            forall<T> { not { T: Foo } }
+        } yields_all[SolverChoice::slg(2, None)] {
+            "substitution [], lifetime constraints []"
+        }
+
+        // However, if we come across a negative goal that exceeds our
+        // size threshold, we have a problem.
+        goal {
+            exists<T> { T = Vec<u32>, not { Vec<Vec<T>>: Foo } }
+        } yields_all[SolverChoice::slg(2, None)] {
+            ""
+        }
+
+        // Same query with larger threshold works fine, though.
+        goal {
+            exists<T> { T = Vec<u32>, not { Vec<Vec<T>>: Foo } }
+        } yields_all[SolverChoice::slg(4, None)] {
+            "substitution [?0 := Vec<u32>], lifetime constraints []"
+        }
+
+        // Here, due to the hypothesis, there does indeed exist a suitable T, `U`.
+        goal {
+            forall<U> { if (U: Foo) { exists<T> { T: Foo } } }
+        } yields_all[SolverChoice::slg(2, None)] {
+            "substitution [?0 := !1_0], lifetime constraints []"
+        }
+    }
+}
+
+#[test]
+fn subgoal_cycle_inhabited() {
+    test! {
+        program {
+            trait Foo { }
+            struct Box<T> { }
+            struct Vec<T> { }
+            struct u32 { }
+            impl<T> Foo for Box<T> where Box<Vec<T>>: Foo { }
+            impl Foo for u32 { }
+        }
+
+        goal {
+            exists<T> { T: Foo }
+        } yields_all[SolverChoice::slg(3, None)] {
+            "substitution [?0 := u32], lifetime constraints []"
+        }
+    }
+}
+
+#[test]
+fn basic_region_constraint_from_positive_impl() {
+    test! {
+        program {
+            trait Foo { }
+            struct Ref<'a, 'b, T> { }
+            struct u32 { }
+            impl<'x, T> Foo for Ref<'x, 'x, T> { }
+        }
+
+        goal {
+            forall<'a, 'b, T> { Ref<'a, 'b, T>: Foo }
+        } yields_all[SolverChoice::slg(3, None)] {
+            "substitution [], lifetime constraints [InEnvironment { environment: Env([]), goal: '!1_1 == '!1_0 }]"
+        }
+    }
+}
+
+#[test]
+#[allow(non_snake_case)]
+fn example_2_1_EWFS() {
+    test! {
+        program {
+            trait Edge<B> { }
+            trait TransitiveClosure<B> { }
+            struct a { }
+            struct b { }
+            struct c { }
+
+            forall<> { a: Edge<b> }
+            forall<> { b: Edge<c> }
+            forall<> { b: Edge<a> }
+            forall<X, Y> { X: TransitiveClosure<Y> if X: Edge<Y> }
+            forall<X, Y, Z> { X: TransitiveClosure<Y> if X: Edge<Z>, Z: TransitiveClosure<Y> }
+        }
+
+        goal {
+            exists<V> { a: TransitiveClosure<V> }
+        } yields_all[SolverChoice::slg(3, None)] {
+            "substitution [?0 := b], lifetime constraints []",
+            "substitution [?0 := c], lifetime constraints []",
+            "substitution [?0 := a], lifetime constraints []"
+        }
+    }
+}
+
+/// Test (along with the other `cached_answers` tests) that the
+/// ordering in which we we encounter clauses doesn't affect the final
+/// set of answers we get. In particular, all of them should get 5
+/// answers, but in Ye Olde Days Of Yore there were sometimes bugs
+/// that came up when replaying tabled answers that led to fewer
+/// answers being produced.
+///
+/// This test is also a test for ANSWER ABSTRACTION: the only reason
+/// we get 5 answers is because of the max size of 2.
+#[test]
+fn cached_answers_1() {
+    test! {
+        program {
+            trait Sour { }
+            struct Lemon { }
+            struct Vinegar { }
+            struct HotSauce<T> { }
+
+            // Use explicit program clauses here rather than traits
+            // and impls to avoid hashmaps and other things that
+            // sometimes alter the final order of the program clauses:
+            forall<> { Lemon: Sour }
+            forall<> { Vinegar: Sour }
+            forall<T> { HotSauce<T>: Sour if T: Sour }
+        }
+
+        goal {
+            exists<T> { T: Sour }
+        } yields_all[SolverChoice::slg(2, None)] {
+            "substitution [?0 := Lemon], lifetime constraints []",
+            "substitution [?0 := Vinegar], lifetime constraints []",
+            "substitution [?0 := HotSauce<Lemon>], lifetime constraints []",
+            "substitution [?0 := HotSauce<Vinegar>], lifetime constraints []",
+            "Ambiguous(for<?U0> { substitution [?0 := HotSauce<^0>], lifetime constraints [] })"
+        }
+    }
+}
+
+/// See `cached_answers_1`.
+#[test]
+fn cached_answers_2() {
+    test! {
+        program {
+            trait Sour { }
+            struct Lemon { }
+            struct Vinegar { }
+            struct HotSauce<T> { }
+
+            forall<T> { HotSauce<T>: Sour if T: Sour }
+            forall<> { Lemon: Sour }
+            forall<> { Vinegar: Sour }
+        }
+
+        goal {
+            exists<T> { T: Sour }
+        } yields_all[SolverChoice::slg(2, None)] {
+            "substitution [?0 := Lemon], lifetime constraints []",
+            "substitution [?0 := Vinegar], lifetime constraints []",
+            "substitution [?0 := HotSauce<Lemon>], lifetime constraints []",
+            "substitution [?0 := HotSauce<Vinegar>], lifetime constraints []",
+            "Ambiguous(for<?U0> { substitution [?0 := HotSauce<^0>], lifetime constraints [] })"
+        }
+    }
+}
+
+/// See `cached_answers_1`.
+#[test]
+fn cached_answers_3() {
+    test! {
+        program {
+            trait Sour { }
+            struct Lemon { }
+            struct Vinegar { }
+            struct HotSauce<T> { }
+
+            forall<> { Lemon: Sour }
+            forall<T> { HotSauce<T>: Sour if T: Sour }
+            forall<> { Vinegar: Sour }
+        }
+
+        goal {
+            exists<T> { T: Sour }
+        } yields_all[SolverChoice::slg(2, None)] {
+            "substitution [?0 := Lemon], lifetime constraints []",
+            "substitution [?0 := HotSauce<Lemon>], lifetime constraints []",
+            "substitution [?0 := Vinegar], lifetime constraints []",
+            "Ambiguous(for<?U0> { substitution [?0 := HotSauce<^0>], lifetime constraints [] })",
+            "substitution [?0 := HotSauce<Vinegar>], lifetime constraints []"
+        }
+    }
+}
+
+#[test]
+fn non_enumerable_traits_direct() {
+    test! {
+        program {
+            struct Foo { }
+            struct Bar { }
+
+            #[non_enumerable]
+            trait NonEnumerable { }
+            impl NonEnumerable for Foo { }
+            impl NonEnumerable for Bar { }
+
+            trait Enumerable { }
+            impl Enumerable for Foo { }
+            impl Enumerable for Bar { }
+        }
+
+        goal {
+            exists<A> { A: NonEnumerable }
+        } yields_first[SolverChoice::slg(3, None)] {
+            "Floundered"
+        }
+
+        goal {
+            exists<A> { A: Enumerable }
+        } yields_all[SolverChoice::slg(3, None)] {
+            "substitution [?0 := Foo], lifetime constraints []",
+            "substitution [?0 := Bar], lifetime constraints []"
+        }
+
+        goal {
+            Foo: NonEnumerable
+        } yields_all[SolverChoice::slg(3, None)] {
+            "substitution [], lifetime constraints []"
+        }
+    }
+}
+
+#[test]
+fn non_enumerable_traits_indirect() {
+    test! {
+        program {
+            struct Foo { }
+            struct Bar { }
+
+            #[non_enumerable]
+            trait NonEnumerable { }
+            impl NonEnumerable for Foo { }
+            impl NonEnumerable for Bar { }
+
+            trait Debug { }
+            impl<T> Debug for T where T: NonEnumerable { }
+        }
+
+        goal {
+            exists<A> { A: Debug }
+        } yields_first[SolverChoice::slg(3, None)] {
+            "Floundered"
+        }
+    }
+}
+
+#[test]
+fn non_enumerable_traits_double() {
+    test! {
+        program {
+            struct Foo { }
+            struct Bar { }
+
+            #[non_enumerable]
+            trait NonEnumerable1 { }
+            impl NonEnumerable1 for Foo { }
+            impl NonEnumerable1 for Bar { }
+
+            #[non_enumerable]
+            trait NonEnumerable2 { }
+            impl NonEnumerable2 for Foo { }
+            impl NonEnumerable2 for Bar { }
+
+            trait Debug { }
+            impl<T> Debug for T where T: NonEnumerable1, T: NonEnumerable2  { }
+        }
+
+        goal {
+            exists<A> { A: Debug }
+        } yields_first[SolverChoice::slg(3, None)] {
+            "Floundered"
+        }
+    }
+}
+
+#[test]
+fn non_enumerable_traits_reorder() {
+    test! {
+        program {
+            struct Foo { }
+            struct Bar { }
+
+            #[non_enumerable]
+            trait NonEnumerable { }
+            impl NonEnumerable for Foo { }
+            impl NonEnumerable for Bar { }
+
+            trait Enumerable { }
+            impl Enumerable for Foo { }
+
+            // In this test, we first try to solve to solve `T:
+            // NonEnumerable` but then we discover it's
+            // non-enumerable, and so we push it off for later. Then
+            // we try to solve the `T: Enumerable` trait.
+
+            trait Debug1 { }
+            impl<T> Debug1 for T where T: Enumerable, T: NonEnumerable { }
+
+            trait Debug2 { }
+            impl<T> Debug2 for T where T: NonEnumerable, T: Enumerable { }
+        }
+
+        goal {
+            exists<A> { A: Debug1 }
+        } yields_all[SolverChoice::slg(3, None)] {
+            "substitution [?0 := Foo], lifetime constraints []"
+        }
+
+
+        goal {
+            exists<A> { A: Debug2 }
+        } yields_all[SolverChoice::slg(3, None)] {
+            "substitution [?0 := Foo], lifetime constraints []"
+        }
+    }
+}
